pyproject.toml
[project]
name = "gobii-platform"
version = "0.1.0"
description = "Platform for browser agents"
requires-python = ">=3.12"

dependencies = [
  "pottery==3.0.1",
  "boto3~=1.38.39",
  "browser-use==0.5.3",
  "openai==1.91.0",
  "pypandoc-binary==1.15",
  "inscriptis==2.6.0",
  "celery~=5.5.2",
  "cryptography~=44.0.3",
  "django-allauth~=65.7.0",
  "django-environ~=0.12.0",
  "django-htmx~=1.23.0",
  "django-storages~=1.14.6",
  "django==5.2",
  "djangorestframework~=3.16.0",
  "celery-redbeat~=2.3.2",
  "django-anymail[postmark]>=13.0",
  "django-anymail[mailgun]>=13.0",
  "drf-nested-routers~=0.94.1",
  "drf-spectacular~=0.28",
  "django-turnstile~=0.1.2",
  "django-waffle~=4.1.0",
  "django-pglock~=1.7.2",
  "google-cloud-storage~=3.1.0",
  "gunicorn==23.0.0",
  "zstandard~=0.23.0",
  "json-schema-to-pydantic~=0.2.5",
  "litellm~=1.73.7",
  "markdown~=3.8",
  "psycopg[binary]~=3.1",
  "pyjwt~=2.10.1",
  "pygments~=2.19.1",
  "python-frontmatter~=1.1.0",
  "ulid-py~=1.1.0",
  "redis~=6.2.0",
  "requests~=2.32.4",
  "segment-analytics-python~=2.3.4",
  "whitenoise[brotli]~=6.9.0",
  "dj-stripe~=2.9.1",
  "tenacity~=8.5.0",
  "opentelemetry-instrumentation~=0.55b1",
  "opentelemetry-distro~=0.55b1",
  "opentelemetry-exporter-otlp~=1.34.1",
  "opentelemetry-api==1.34.1",
  "opentelemetry-exporter-otlp-proto-common==1.34.1",
  "opentelemetry-exporter-otlp-proto-grpc==1.34.1",
  "opentelemetry-exporter-otlp-proto-http==1.34.1",
  "opentelemetry-instrumentation-alephalpha==0.40.11",
  "opentelemetry-instrumentation-anthropic==0.40.11",
  "opentelemetry-instrumentation-asgi==0.55b1",
  "opentelemetry-instrumentation-asyncio==0.55b1",
  "opentelemetry-instrumentation-bedrock==0.40.11",
  "opentelemetry-instrumentation-boto3sqs==0.55b1",
  "opentelemetry-instrumentation-botocore==0.55b1",
  "opentelemetry-instrumentation-celery==0.55b1",
  "opentelemetry-instrumentation-chromadb==0.40.11",
  "opentelemetry-instrumentation-click==0.55b1",
  "opentelemetry-instrumentation-cohere==0.40.11",
  "opentelemetry-instrumentation-crewai==0.40.11",
  "opentelemetry-instrumentation-dbapi==0.55b1",
  "opentelemetry-instrumentation-django==0.55b1",
  "opentelemetry-instrumentation-google-generativeai==0.40.11",
  "opentelemetry-instrumentation-groq==0.40.11",
  "opentelemetry-instrumentation-grpc==0.55b1",
  "opentelemetry-instrumentation-haystack==0.40.11",
  "opentelemetry-instrumentation-httpx==0.55b1",
  "opentelemetry-instrumentation-lancedb==0.40.11",
  "opentelemetry-instrumentation-langchain==0.40.11",
  "opentelemetry-instrumentation-llamaindex==0.40.11",
  "opentelemetry-instrumentation-logging==0.55b1",
  "opentelemetry-instrumentation-marqo==0.40.11",
  "opentelemetry-instrumentation-mcp==0.40.11",
  "opentelemetry-instrumentation-milvus==0.40.11",
  "opentelemetry-instrumentation-mistralai==0.40.11",
  "opentelemetry-instrumentation-ollama==0.40.11",
  "opentelemetry-instrumentation-openai==0.40.11",
  "opentelemetry-instrumentation-openai-v2==2.1b0",
  "opentelemetry-instrumentation-pinecone==0.40.11",
  "opentelemetry-instrumentation-psycopg==0.55b1",
  "opentelemetry-instrumentation-psycopg2==0.55b1",
  "opentelemetry-instrumentation-qdrant==0.40.11",
  "opentelemetry-instrumentation-redis==0.55b1",
  "opentelemetry-instrumentation-replicate==0.40.11",
  "opentelemetry-instrumentation-requests==0.55b1",
  "opentelemetry-instrumentation-sagemaker==0.40.11",
  "opentelemetry-instrumentation-sqlalchemy==0.55b1",
  "opentelemetry-instrumentation-sqlite3==0.55b1",
  "opentelemetry-instrumentation-system-metrics==0.55b1",
  "opentelemetry-instrumentation-threading==0.55b1",
  "opentelemetry-instrumentation-together==0.40.11",
  "opentelemetry-instrumentation-tortoiseorm==0.55b1",
  "opentelemetry-instrumentation-transformers==0.40.11",
  "opentelemetry-instrumentation-urllib==0.55b1",
  "opentelemetry-instrumentation-urllib3==0.55b1",
  "opentelemetry-instrumentation-vertexai==0.40.11",
  "opentelemetry-instrumentation-watsonx==0.40.11",
  "opentelemetry-instrumentation-weaviate==0.40.11",
  "opentelemetry-instrumentation-wsgi==0.55b1",
  "opentelemetry-propagator-aws-xray==1.0.2",
  "opentelemetry-proto==1.34.1",
  "opentelemetry-sdk==1.34.1",
  "opentelemetry-semantic-conventions==0.55b1",
  "opentelemetry-semantic-conventions-ai==0.4.9",
  "opentelemetry-util-http==0.55b1",
  "mail-parser==4.1.4",
  "exa-py==1.14.16",
  # Email and SMS integration
  "aiohttp-retry==2.9.1",
  "twilio==9.6.3",
  "phonenumbers==9.0.9",
  # Url matching
  "urlextract==1.9.0",
  "fastmcp>=2.11.3",
]

[build-system]
requires = ["uv", "setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where   = ["."]
include = ["config", "console*", "api*", "pages*"]
exclude = ["*/tests*", "*/static*", "*/templates*"]
